requirements.txt
Django>=4.2
djangorestframework
python-decouple
django-filter
djangorestframework-simplejwt
drf-spectacular


